incomporate koenkk's suggestion to isolate the changes into one file
